refactor: extract decorator route runtime
Move selection, reconciliation, normalization, materialization, and
identity mechanics into a cycle-safe subsystem. Include decorator set
cardinality observation. Keep mixed nested coordinators in the facade.

All 26 definitions are source- and AST-identical to origin/main.
Retain one selection type identity.
Preserve canonical excluded-field filtering.

Validation: 858 tests passed on Pydantic 2.13.4 (95.85% coverage).
Validation: 858 tests passed on Pydantic 2.12.3 (95.78% coverage).
Validation: 392 focused tests and 43 independent boundaries passed.
Validation: Ruff, ty, strict mypy, Vulture, and strict docs passed.
Validation: immutable v0.3.0 contract passed.

Closes #116
Refs #94
