+++ src/pydantic_versions/_runtime.py
@@ -1,31 +1,39 @@
 from __future__ import annotations
 
-from collections.abc import Callable, Mapping
-from dataclasses import dataclass
+from collections.abc import Mapping
 from functools import partial
-from types import UnionType
 from typing import (
     TYPE_CHECKING,
     Any,
     Literal,
-    Union,
-    get_args,
-    get_origin,
 )
 
 from pydantic import BaseModel
 
-from pydantic_versions._compiler import (
-    _CompiledDecoratorNestedFamily,
-    _CompiledFamily,
-    _CompiledVersion,
+from pydantic_versions._compiler import _CompiledFamily
+from pydantic_versions._runtime_decorators import (
+    _decorator_collection_kind,
+    _decorator_selections_child_first,
+    _decorator_selections_parent_first,
+    _decorator_set_cardinalities,
+    _DecoratorRouteSelection,
+    _materialize_selected_decorator_models,
+    _normalize_decorator_value,
+    _normalize_selected_decorator_payloads,
+    _payload_at_location,
+    _reconcile_decorator_selections,
+    _refresh_decorator_selection_identities,
+    _refresh_decorator_selection_identity,
+    _select_decorator_routes,
+    _serialized_decorator_selection_location,
+    _transform_payload_at_location,
+    _walk_decorator_payload_candidates,
 )
 from pydantic_versions._runtime_nested import (
     _apply_nested_family_migrations,
     _convert_nested_family_payload,
     _declared_nested_version_metadata,
     _nested_family_collection_kind,
-    _normalize_validated_explicit_nested_payloads,
     _preflight_nested_version_metadata,
     _preflight_validation_route,
     _project_nested_family_payload,
@@ -41,9 +49,7 @@ from pydantic_versions._runtime_payload import (
     _explicit_runtime_body_model,
     _extract_declared_fields,
     _extract_preflight_fields,
-    _matching_declared_annotation,
     _runtime_nested_structural_occurrences,
-    _strip_annotated,
 )
 from pydantic_versions._runtime_render import (
     _copy_render_input,
@@ -78,17 +84,6 @@ if TYPE_CHECKING:
     from pydantic_versions.family import SchemaFamily
 
 
-@dataclass
-class _DecoratorRouteSelection:
-    route: _CompiledDecoratorNestedFamily
-    location: tuple[str | int, ...]
-    relative_location: tuple[str | int, ...]
-    site_routes: tuple[_CompiledDecoratorNestedFamily, ...]
-    label: str
-    parent: _DecoratorRouteSelection | None = None
-    value_identity: int | None = None
-
-
 _CONTRACT_FIELD_OMISSION_OPTIONS = frozenset(
     {
         "exclude",
@@ -111,151 +106,6 @@ _SUPPORTED_MODEL_DUMP_OPTIONS = frozenset(
         "warnings",
     }
 )
-
-
-def _select_decorator_routes(
-    value: BaseModel,
-    *,
-    compiled: _CompiledFamily,
-    parent_label: str,
-    source_version: _CompiledVersion | None,
-    location_prefix: tuple[str | int, ...] = (),
-    parent_selection: _DecoratorRouteSelection | None = None,
-) -> tuple[_DecoratorRouteSelection, ...]:
-    if not compiled.decorator_nested:
-        return ()
-    root_names = {
-        route.path[0]: (
-            route.path[0]
-            if source_version is None
-            else source_version.projection.field(route.path[0]).version_name
-        )
-        for route in compiled.decorator_nested
-    }
-    selected: list[_DecoratorRouteSelection] = []
-    for route in compiled.decorator_nested:
-        if root_names[route.path[0]] is None:
-            continue
-        for nested_value, location in _walk_authoritative_decorator_route(
-            value,
-            annotation=type(value),
-            route=route,
-            root_names=root_names,
-        ):
-            expected = (route.family.model, route.family.model_for(parent_label))
-            if isinstance(nested_value, expected):
-                site = _decorator_dispatch_site(route)
-                selection = _DecoratorRouteSelection(
-                    route=route,
-                    location=(*location_prefix, *location),
-                    relative_location=location,
-                    site_routes=tuple(
-                        candidate
-                        for candidate in compiled.decorator_nested
-                        if _decorator_dispatch_site(candidate) == site
-                    ),
-                    label=route.child_label(parent_label),
-                    parent=parent_selection,
-                )
-                selected.append(selection)
-                child_compiled = route.family._compiled_family()
-                child_source = (
-                    None
-                    if isinstance(nested_value, route.family.model)
-                    else child_compiled.version(selection.label)
-                )
-                selected.extend(
-                    _select_decorator_routes(
-                        nested_value,
-                        compiled=child_compiled,
-                        parent_label=selection.label,
-                        source_version=child_source,
-                        location_prefix=selection.location,
-                        parent_selection=selection,
-                    )
-                )
-    return tuple(selected)
-
-
-def _walk_authoritative_decorator_route(
-    value: Any,
-    *,
-    annotation: Any,
-    route: _CompiledDecoratorNestedFamily,
-    root_names: Mapping[str, str | None],
-) -> tuple[tuple[Any, tuple[str | int, ...]], ...]:
-    states: list[tuple[Any, Any, tuple[str | int, ...]]] = [(value, annotation, ())]
-    for step_index, step in enumerate(route.traversal):
-        next_states: list[tuple[Any, Any, tuple[str | int, ...]]] = []
-        for current, current_annotation, location in states:
-            normalized_annotation = _strip_annotated(current_annotation)
-            if step.kind == "field":
-                if not isinstance(current, BaseModel):
-                    continue
-                actual_name = root_names.get(step.value) if step_index == 0 else step.value
-                if actual_name is None:
-                    continue
-                model = type(current)
-                field_info = model.model_fields.get(actual_name)
-                if field_info is None or actual_name not in current.__dict__:
-                    continue
-                next_states.append(
-                    (
-                        current.__dict__[actual_name],
-                        field_info.annotation,
-                        (*location, step.value),
-                    )
-                )
-                continue
-            if step.kind == "union_arm":
-                if get_origin(normalized_annotation) not in (Union, UnionType):
-                    continue
-                arguments = get_args(normalized_annotation)
-                ordinal = int(step.value)
-                if ordinal >= len(arguments):
-                    continue
-                selected = _matching_declared_annotation(normalized_annotation, current)
-                candidate = _strip_annotated(arguments[ordinal])
-                if _strip_annotated(selected) != candidate:
-                    continue
-                next_states.append((current, arguments[ordinal], location))
-                continue
-            if step.kind == "each":
-                if not isinstance(current, list | tuple | set | frozenset):
-                    continue
-                arguments = get_args(normalized_annotation)
-                if not arguments:
-                    continue
-                item_annotation = arguments[0]
-                next_states.extend(
-                    (item, item_annotation, (*location, ordinal))
-                    for ordinal, item in enumerate(current)
-                )
-                continue
-            if step.kind == "tuple_index":
-                if not isinstance(current, tuple | list):
-                    continue
-                arguments = get_args(normalized_annotation)
-                ordinal = int(step.value)
-                if ordinal >= len(current) or ordinal >= len(arguments):
-                    continue
-                next_states.append((current[ordinal], arguments[ordinal], (*location, ordinal)))
-                continue
-            if step.kind == "mapping_values":
-                if not isinstance(current, Mapping):
-                    continue
-                arguments = get_args(normalized_annotation)
-                if len(arguments) != 2:
-                    continue
-                next_states.extend(
-                    (item, arguments[1], (*location, key))
-                    for key, item in current.items()
-                    if isinstance(key, str | int)
-                )
-        states = next_states
-        if not states:
-            break
-    return tuple((item, location) for item, _, location in states)
 
 
 def _validate_family[T: BaseModel](
@@ -750,75 +600,6 @@ def _prune_serialized_decorator_metadata(
         )
 
 
-def _serialized_decorator_selection_location(
-    *,
-    compiled: _CompiledFamily,
-    parent_label: str,
-    selection: _DecoratorRouteSelection,
-    by_alias: Any,
-) -> tuple[str | int, ...] | None:
-    if selection.parent is None:
-        prefix: tuple[str | int, ...] = ()
-        owner_compiled = compiled
-        owner_label = parent_label
-    else:
-        parent_location = _serialized_decorator_selection_location(
-            compiled=compiled,
-            parent_label=parent_label,
-            selection=selection.parent,
-            by_alias=by_alias,
-        )
-        if parent_location is None:
-            return None
-        prefix = parent_location
-        owner_compiled = selection.parent.route.family._compiled_family()
-        owner_label = selection.parent.label
-
-    owner_target = owner_compiled.version(owner_label)
-    annotation: Any = owner_target.model
-    relative = iter(selection.relative_location)
-    serialized: list[str | int] = list(prefix)
-    for step_index, step in enumerate(selection.route.traversal):
-        normalized = _strip_annotated(annotation)
-        if step.kind == "field":
-            next(relative)
-            field_name = step.value
-            if step_index == 0:
-                projected_name = owner_target.projection.field(field_name).version_name
-                if projected_name is None:
-                    return None
-                field_name = projected_name
-            field_info = normalized.model_fields[field_name]
-            serialized.append(_serialized_field_name(normalized, field_name, by_alias=by_alias))
-            annotation = field_info.annotation
-            continue
-        if step.kind == "union_arm":
-            arguments = get_args(normalized)
-            ordinal = int(step.value)
-            annotation = arguments[ordinal]
-            continue
-        if step.kind == "each":
-            occurrence = next(relative)
-            arguments = get_args(normalized)
-            serialized.append(occurrence)
-            annotation = arguments[0]
-            continue
-        if step.kind == "tuple_index":
-            occurrence = next(relative)
-            ordinal = int(step.value)
-            arguments = get_args(normalized)
-            serialized.append(occurrence)
-            annotation = arguments[ordinal]
-            continue
-        if step.kind == "mapping_values":
-            occurrence = next(relative)
-            arguments = get_args(normalized)
-            serialized.append(occurrence)
-            annotation = arguments[1]
-            continue
-    return tuple(serialized)
-
-
 def _validated_current_render_payload[T: BaseModel](
     *,
     family: SchemaFamily[T],
@@ -892,55 +673,6 @@ def _validated_current_render_payload[T: BaseModel](
     )
     _refresh_decorator_selection_identities(payload, selections)
     return payload, selections
-
-
-def _normalize_selected_decorator_payloads(
-    *,
-    payload: dict[str, Any],
-    selections: tuple[_DecoratorRouteSelection, ...],
-    parent_label: str,
-) -> dict[str, Any]:
-    normalized = payload
-    # Historical ancestor field names must become canonical before a
-    # descendant's canonical location is reachable.  Conversion and final
-    # projection remain child-first, but input-name normalization is the
-    # inverse traversal.
-    for selection in _decorator_selections_parent_first(selections):
-        child_compiled = selection.route.family._compiled_family()
-        child_label = selection.route.child_label(parent_label)
-        normalized = _transform_payload_at_location(
-            normalized,
-            location=selection.location,
-            transform=partial(
-                _normalize_decorator_value,
-                compiled=child_compiled,
-                child_label=child_label,
-            ),
-        )
-        _refresh_decorator_selection_identity(normalized, selection)
-    return normalized
-
-
-def _normalize_decorator_value(
-    value: Any,
-    *,
-    compiled: _CompiledFamily,
-    child_label: str,
-) -> Any:
-    if isinstance(value, BaseModel):
-        value = _extract_declared_fields(value)
-    if not isinstance(value, Mapping):
-        return value
-    normalized = _to_current_names(
-        compiled,
-        compiled.version(child_label),
-        dict(value),
-    )
-    return _normalize_validated_explicit_nested_payloads(
-        payload=normalized,
-        compiled=compiled,
-        parent_label=child_label,
-    )
 
 
 def _apply_selected_decorator_migrations(
@@ -1033,611 +765,6 @@ def _project_decorator_value(
     )
 
 
-def _materialize_selected_decorator_models(
-    *,
-    payload: dict[str, Any],
-    selections: tuple[_DecoratorRouteSelection, ...],
-) -> dict[str, Any]:
-    materialized = payload
-    for selection in _decorator_selections_child_first(selections):
-        materialized = _transform_payload_at_location(
-            materialized,
-            location=selection.location,
-            transform=partial(
-                _materialize_decorator_model,
-                model=selection.route.family.model,
-            ),
-        )
-        _refresh_decorator_selection_identity(materialized, selection)
-    return materialized
-
-
-def _materialize_decorator_model(value: Any, *, model: type[BaseModel]) -> Any:
-    if type(value) is model:
-        return value
-    if not isinstance(value, Mapping):
-        return value
-    validation_input = _current_validation_input(model, dict(value))
-    if model.model_config.get("revalidate_instances") == "always":
-        # Preserve the authoritative branch as an exact instance, but let the
-        # final parent validation perform the one real validation required by
-        # the model's explicit revalidation policy.
-        return model.model_construct(**validation_input)
-    return model.model_validate(
-        validation_input,
-        by_name=True,
-    )
-
-
-def _decorator_collection_kind(
-    route: _CompiledDecoratorNestedFamily,
-) -> Literal["list", "tuple", "set", "frozenset"] | None:
-    if route.collection_kind == "mapping":
-        return None
-    return route.collection_kind
-
-
-def _transform_payload_at_location(
-    payload: Any,
-    *,
-    location: tuple[str | int, ...],
-    transform: Callable[[Any], Any],
-) -> Any:
-    if not location:
-        return transform(payload)
-    head, *remaining = location
-    tail = tuple(remaining)
-    if isinstance(payload, BaseModel):
-        return _transform_payload_at_location(
-            _extract_declared_fields(payload),
-            location=location,
-            transform=transform,
-        )
-    if isinstance(payload, Mapping):
-        if head not in payload:
-            return payload
-        current = payload[head]
-        transformed = _transform_payload_at_location(
-            current,
-            location=tail,
-            transform=transform,
-        )
-        if transformed is current:
-            return payload
-        copied = dict(payload)
-        copied[head] = transformed
-        return copied
-    if isinstance(payload, list | tuple) and isinstance(head, int):
-        if head < 0 or head >= len(payload):
-            return payload
-        current = payload[head]
-        transformed = _transform_payload_at_location(
-            current,
-            location=tail,
-            transform=transform,
-        )
-        if transformed is current:
-            return payload
-        copied_items = list(payload)
-        copied_items[head] = transformed
-        return copied_items if isinstance(payload, list) else tuple(copied_items)
-    return payload
-
-
-def _payload_at_location(
-    payload: Any,
-    location: tuple[str | int, ...],
-) -> tuple[bool, Any]:
-    current = payload
-    for part in location:
-        if isinstance(current, Mapping):
-            if part not in current:
-                return False, None
-            current = current[part]
-            continue
-        if isinstance(current, list | tuple) and isinstance(part, int):
-            if part < 0 or part >= len(current):
-                return False, None
-            current = current[part]
-            continue
-        return False, None
-    return True, current
-
-
-def _refresh_decorator_selection_identities(
-    payload: Any,
-    selections: tuple[_DecoratorRouteSelection, ...],
-) -> None:
-    for selection in selections:
-        _refresh_decorator_selection_identity(payload, selection)
-
-
-def _refresh_decorator_selection_identity(
-    payload: Any,
-    selection: _DecoratorRouteSelection,
-) -> None:
-    found, value = _payload_at_location(payload, selection.location)
-    selection.value_identity = (
-        id(value) if found and isinstance(value, Mapping | BaseModel) else None
-    )
-
-
-def _decorator_selection_depth(selection: _DecoratorRouteSelection) -> int:
-    depth = 0
-    parent = selection.parent
-    while parent is not None:
-        depth += 1
-        parent = parent.parent
-    return depth
-
-
-def _decorator_selections_child_first(
-    selections: tuple[_DecoratorRouteSelection, ...],
-) -> tuple[_DecoratorRouteSelection, ...]:
-    return tuple(
-        sorted(
-            selections,
-            key=_decorator_selection_depth,
-            reverse=True,
-        )
-    )
-
-
-def _decorator_selections_parent_first(
-    selections: tuple[_DecoratorRouteSelection, ...],
-) -> tuple[_DecoratorRouteSelection, ...]:
-    return tuple(sorted(selections, key=_decorator_selection_depth))
-
-
-def _decorator_dispatch_site(
-    route: _CompiledDecoratorNestedFamily,
-) -> tuple[tuple[str, str], ...]:
-    return tuple(
-        (step.kind, "*") if step.kind == "union_arm" else (step.kind, step.value)
-        for step in route.traversal
-    )
-
-
-def _walk_decorator_payload_candidates(
-    payload: Any,
-    route: _CompiledDecoratorNestedFamily,
-) -> tuple[tuple[Any, tuple[str | int, ...]], ...]:
-    states: list[tuple[Any, tuple[str | int, ...]]] = [(payload, ())]
-    for step in route.traversal:
-        next_states: list[tuple[Any, tuple[str | int, ...]]] = []
-        for current, location in states:
-            if step.kind == "field":
-                if isinstance(current, BaseModel):
-                    if step.value not in current.__dict__:
-                        continue
-                    next_states.append((current.__dict__[step.value], (*location, step.value)))
-                elif isinstance(current, Mapping) and step.value in current:
-                    next_states.append((current[step.value], (*location, step.value)))
-                continue
-            if step.kind == "union_arm":
-                next_states.append((current, location))
-                continue
-            if step.kind == "each":
-                if not isinstance(current, list | tuple | set | frozenset):
-                    continue
-                next_states.extend(
-                    (item, (*location, ordinal)) for ordinal, item in enumerate(current)
-                )
-                continue
-            if step.kind == "tuple_index":
-                if not isinstance(current, list | tuple):
-                    continue
-                ordinal = int(step.value)
-                if ordinal < len(current):
-                    next_states.append((current[ordinal], (*location, ordinal)))
-                continue
-            if step.kind == "mapping_values":
-                if not isinstance(current, Mapping):
-                    continue
-                if any(not isinstance(key, str) for key in current):
-                    msg = (
-                        "Parent migration introduced a non-string decorator nested "
-                        f"mapping key at path {route.path!r}"
-                    )
-                    raise InvalidMigrationError(msg)
-                next_states.extend(
-                    (item, (*location, key))
-                    for key, item in current.items()
-                    if isinstance(key, str)
-                )
-        states = next_states
-        if not states:
-            break
-    return tuple(
-        (value, location) for value, location in states if isinstance(value, Mapping | BaseModel)
-    )
-
-
-def _route_has_mapping_anchor(route: _CompiledDecoratorNestedFamily) -> bool:
-    if not any(step.kind == "mapping_values" for step in route.traversal):
-        return False
-    # A mapping key is only an occurrence anchor when the complete route to it
-    # is structurally stable.  A dynamic collection on either side of the
-    # mapping can reorder while retaining the same inner key, so anchoring the
-    # new absolute location would silently transfer a union branch identity.
-    return not any(step.kind == "each" for step in route.traversal)
-
-
-def _reconcile_decorator_selections(
-    *,
-    payload: dict[str, Any],
-    selections: tuple[_DecoratorRouteSelection, ...],
-    compiled: _CompiledFamily,
-    discover_new: bool,
-) -> tuple[_DecoratorRouteSelection, ...]:
-    if not selections:
-        if not discover_new:
-            return selections
-        return _discover_typed_decorator_selections(
-            payload=payload,
-            compiled=compiled,
-            selections=selections,
-        )
-
-    selected_by_site: dict[
-        tuple[int, tuple[tuple[str, str], ...]],
-        list[_DecoratorRouteSelection],
-    ] = {}
-    for selection in selections:
-        site = _decorator_dispatch_site(selection.route)
-        selected_by_site.setdefault((id(selection.parent), site), []).append(selection)
-
-    identity_sites = {
-        selection.value_identity: site_key
-        for site_key, site_selections in selected_by_site.items()
-        for selection in site_selections
-        if selection.value_identity is not None
-    }
-    replaced_owner_ids: set[int] = set()
-    groups = tuple(
-        sorted(
-            selected_by_site.items(),
-            key=lambda item: _decorator_selection_depth(item[1][0]),
-        )
-    )
-    for site_key, previous in groups:
-        if any(
-            _selection_has_replaced_ancestor(selection, replaced_owner_ids)
-            for selection in previous
-        ):
-            continue
-        declarations = previous[0].site_routes
-        parent = previous[0].parent
-        if parent is None:
-            base_payload = payload
-            location_prefix: tuple[str | int, ...] = ()
-        else:
-            found, base_payload = _payload_at_location(payload, parent.location)
-            if not found:
-                msg = (
-                    "Parent migration removed a decorator nested owner occurrence at "
-                    f"path {parent.route.path!r}"
-                )
-                raise InvalidMigrationError(msg)
-            location_prefix = parent.location
-        candidates = [
-            (value, (*location_prefix, *relative_location), relative_location)
-            for value, relative_location in _walk_decorator_payload_candidates(
-                base_payload,
-                declarations[0],
-            )
-        ]
-        if any(
-            id(value) in identity_sites and identity_sites[id(value)] != site_key
-            for value, _, _ in candidates
-        ):
-            msg = (
-                "Parent migration moved a decorator nested occurrence across dispatch "
-                f"sites at path {declarations[0].path!r}"
-            )
-            raise InvalidMigrationError(msg)
-        if len(candidates) < len(previous):
-            msg = (
-                "Parent migration changed decorator nested occurrence cardinality at "
-                f"path {declarations[0].path!r}; typed replacements must remain one-to-one"
-            )
-            raise InvalidMigrationError(msg)
-
-        remaining_candidates = list(enumerate(candidates))
-        remaining_previous = list(previous)
-
-        # Object identity is the authoritative occurrence token. It permits a
-        # heterogeneous list to reorder without rediscovering its union branch.
-        for selection in tuple(remaining_previous):
-            if selection.value_identity is None:
-                continue
-            matches = [
-                (index, candidate)
-                for index, candidate in remaining_candidates
-                if id(candidate[0]) == selection.value_identity
-            ]
-            if len(matches) > 1:
-                msg = (
-                    "Parent migration reused one decorator nested occurrence at "
-                    f"path {selection.route.path!r}"
-                )
-                raise InvalidMigrationError(msg)
-            if not matches:
-                continue
-            candidate_index, (value, location, relative_location) = matches[0]
-            selection.location = location
-            selection.relative_location = relative_location
-            selection.value_identity = id(value)
-            remaining_previous.remove(selection)
-            remaining_candidates = [
-                item for item in remaining_candidates if item[0] != candidate_index
-            ]
-
-        # An exact current-model instance is an explicit, authoritative branch
-        # replacement and may establish a new route without validation.
-        for candidate_index, (value, location, relative_location) in tuple(remaining_candidates):
-            if not remaining_previous:
-                break
-            if not isinstance(value, BaseModel):
-                continue
-            matching = tuple(route for route in declarations if type(value) is route.family.model)
-            if len(matching) != 1:
-                msg = (
-                    "Parent migration supplied a typed decorator nested replacement "
-                    f"with no unique family at path {declarations[0].path!r}"
-                )
-                raise InvalidMigrationError(msg)
-            route = matching[0]
-            replaced = next(
-                (selection for selection in remaining_previous if selection.location == location),
-                remaining_previous[0],
-            )
-            replaced.route = route
-            replaced.location = location
-            replaced.relative_location = relative_location
-            replaced.label = route.family.current_version
-            replaced.value_identity = id(value)
-            replaced_owner_ids.add(id(replaced))
-            remaining_previous.remove(replaced)
-            remaining_candidates = [
-                item for item in remaining_candidates if item[0] != candidate_index
-            ]
-
-        if not remaining_previous:
-            if remaining_candidates and not all(
-                len(tuple(route for route in declarations if type(value) is route.family.model))
-                == 1
-                for _, (value, _, _) in remaining_candidates
-            ):
-                msg = (
-                    "Parent migration changed decorator nested occurrence cardinality at "
-                    f"path {declarations[0].path!r}; new occurrences must be exact "
-                    "current-model instances"
-                )
-                raise InvalidMigrationError(msg)
-            continue
-
-        # Dictionary keys are stable anchors even if a callback copies the
-        # nested value. List/set positions and bare union fields are not.
-        if all(_route_has_mapping_anchor(selection.route) for selection in remaining_previous):
-            anchored = {selection.location: selection for selection in remaining_previous}
-            if len(anchored) == len(remaining_previous) and all(
-                location in anchored for _, (_, location, _) in remaining_candidates
-            ):
-                for _, (value, location, relative_location) in remaining_candidates:
-                    selection = anchored[location]
-                    selection.relative_location = relative_location
-                    selection.value_identity = id(value)
-                continue
-
-        remaining_families = {selection.route.family for selection in remaining_previous}
-        declared_families = {route.family for route in declarations}
-        if len(remaining_families) == 1 and len(declared_families) == 1:
-            if len(remaining_candidates) != len(remaining_previous):
-                msg = (
-                    "Parent migration changed decorator nested occurrence cardinality at "
-                    f"path {declarations[0].path!r}; typed replacements must remain "
-                    "one-to-one"
-                )
-                raise InvalidMigrationError(msg)
-            for selection, (_, (value, location, relative_location)) in zip(
-                remaining_previous,
-                remaining_candidates,
-                strict=True,
-            ):
-                selection.location = location
-                selection.relative_location = relative_location
-                selection.value_identity = id(value)
-            continue
-
-        msg = (
-            "Parent migration replaced overlapping decorator nested union values "
-            f"with ambiguous raw mappings at path {declarations[0].path!r}; use exact "
-            "current-model instances to establish branch identity"
-        )
-        raise InvalidMigrationError(msg)
-
-    retained = tuple(
-        selection
-        for selection in selections
-        if not _selection_has_replaced_ancestor(selection, replaced_owner_ids)
-    )
-    _validate_unique_decorator_selection_identities(retained)
-    if not discover_new:
-        return retained
-    discovered = _discover_typed_decorator_selections(
-        payload=payload,
-        compiled=compiled,
-        selections=retained,
-    )
-    _validate_unique_decorator_selection_identities(discovered)
-    return discovered
-
-
-def _selection_has_replaced_ancestor(
-    selection: _DecoratorRouteSelection,
-    replaced_owner_ids: set[int],
-) -> bool:
-    parent = selection.parent
-    while parent is not None:
-        if id(parent) in replaced_owner_ids:
-            return True
-        parent = parent.parent
-    return False
-
-
-def _validate_unique_decorator_selection_identities(
-    selections: tuple[_DecoratorRouteSelection, ...],
-) -> None:
-    locations_by_identity: dict[int, tuple[str | int, ...]] = {}
-    for selection in selections:
-        identity = selection.value_identity
-        if identity is None:
-            continue
-        prior_location = locations_by_identity.get(identity)
-        if prior_location is not None and prior_location != selection.location:
-            msg = (
-                "Parent migration reused one decorator nested occurrence at "
-                f"path {selection.route.path!r}"
-            )
-            raise InvalidMigrationError(msg)
-        locations_by_identity[identity] = selection.location
-
-
-def _discover_typed_decorator_selections(
-    *,
-    payload: dict[str, Any],
-    compiled: _CompiledFamily,
-    selections: tuple[_DecoratorRouteSelection, ...],
-) -> tuple[_DecoratorRouteSelection, ...]:
-    discovered = list(selections)
-
-    def visit_owner(
-        owner_payload: Any,
-        owner_compiled: _CompiledFamily,
-        *,
-        location_prefix: tuple[str | int, ...],
-        parent: _DecoratorRouteSelection | None,
-    ) -> None:
-        sites: dict[
-            tuple[tuple[str, str], ...],
-            list[_CompiledDecoratorNestedFamily],
-        ] = {}
-        for route in owner_compiled.decorator_nested:
-            sites.setdefault(_decorator_dispatch_site(route), []).append(route)
-
-        for routes in sites.values():
-            declarations = tuple(routes)
-            for value, relative_location in _walk_decorator_payload_candidates(
-                owner_payload,
-                declarations[0],
-            ):
-                location = (*location_prefix, *relative_location)
-                if any(
-                    selection.parent is parent and selection.location == location
-                    for selection in discovered
-                ):
-                    continue
-                matches = tuple(
-                    route for route in declarations if type(value) is route.family.model
-                )
-                if len(matches) != 1:
-                    if isinstance(value, Mapping) and not _route_has_non_child_mapping_arm(
-                        owner_compiled.model,
-                        declarations[0],
-                    ):
-                        msg = (
-                            "Parent migration introduced an untyped decorator nested "
-                            f"mapping at path {declarations[0].path!r}; use an exact "
-                            "current-model instance to establish branch identity"
-                        )
-                        raise InvalidMigrationError(msg)
-                    continue
-                route = matches[0]
-                if any(
-                    selection.value_identity == id(value) and selection.location != location
-                    for selection in discovered
-                ):
-                    msg = (
-                        "Parent migration reused one decorator nested occurrence at "
-                        f"path {route.path!r}"
-                    )
-                    raise InvalidMigrationError(msg)
-                discovered.append(
-                    _DecoratorRouteSelection(
-                        route=route,
-                        location=location,
-                        relative_location=relative_location,
-                        site_routes=declarations,
-                        label=route.family.current_version,
-                        parent=parent,
-                        value_identity=id(value),
-                    )
-                )
-
-        children = tuple(selection for selection in discovered if selection.parent is parent)
-        for child in children:
-            found, child_payload = _payload_at_location(payload, child.location)
-            if not found:
-                continue
-            visit_owner(
-                child_payload,
-                child.route.family._compiled_family(),
-                location_prefix=child.location,
-                parent=child,
-            )
-
-    visit_owner(payload, compiled, location_prefix=(), parent=None)
-    return tuple(discovered)
-
-
-def _route_has_non_child_mapping_arm(
-    model: type[BaseModel],
-    route: _CompiledDecoratorNestedFamily,
-) -> bool:
-    annotation: Any = model
-    for step in route.traversal:
-        normalized = _strip_annotated(annotation)
-        if step.kind == "field":
-            annotation = normalized.model_fields[step.value].annotation
-            continue
-        if step.kind == "union_arm":
-            arguments = get_args(normalized)
-            ordinal = int(step.value)
-            if any(
-                _annotation_accepts_raw_mapping(argument)
-                for index, argument in enumerate(arguments)
-                if index != ordinal
-            ):
-                return True
-            annotation = arguments[ordinal]
-            continue
-        arguments = get_args(normalized)
-        if step.kind in ("each", "mapping_values"):
-            annotation = arguments[1] if step.kind == "mapping_values" else arguments[0]
-            continue
-        if step.kind == "tuple_index":
-            ordinal = int(step.value)
-            annotation = arguments[ordinal]
-    return False
-
-
-def _annotation_accepts_raw_mapping(annotation: Any) -> bool:
-    normalized = _strip_annotated(annotation)
-    if normalized is Any:
-        return True
-    origin = get_origin(normalized)
-    if origin in (Union, UnionType):
-        return any(_annotation_accepts_raw_mapping(item) for item in get_args(normalized))
-    runtime_type = origin if isinstance(origin, type) else normalized
-    if not isinstance(runtime_type, type):
-        return False
-    try:
-        return issubclass(dict, runtime_type)
-    except TypeError:
-        return False
-
-
 def _preflight_selected_decorator_version_metadata(
     *,
     payload: Any,
@@ -1717,46 +844,6 @@ def _validate_nested_collection_cardinality(
         compiled=compiled,
         parent_label=parent_label,
     )
-
-
-def _decorator_set_cardinalities(
-    payload: Any,
-    route: _CompiledDecoratorNestedFamily,
-) -> dict[int, tuple[int, ...]]:
-    states: list[Any] = [payload]
-    cardinalities: dict[int, list[int]] = {}
-    for step_index, step in enumerate(route.traversal):
-        next_states: list[Any] = []
-        for current in states:
-            if step.kind == "field":
-                if isinstance(current, BaseModel) and step.value in current.__dict__:
-                    next_states.append(current.__dict__[step.value])
-                elif isinstance(current, Mapping) and step.value in current:
-                    next_states.append(current[step.value])
-                continue
-            if step.kind == "union_arm":
-                next_states.append(current)
-                continue
-            if step.kind == "each":
-                if not isinstance(current, list | tuple | set | frozenset):
-                    continue
-                if step.value in ("set", "frozenset"):
-                    cardinalities.setdefault(step_index, []).append(len(current))
-                next_states.extend(current)
-                continue
-            if step.kind == "tuple_index":
-                if not isinstance(current, list | tuple):
-                    continue
-                ordinal = int(step.value)
-                if ordinal < len(current):
-                    next_states.append(current[ordinal])
-                continue
-            if step.kind == "mapping_values" and isinstance(current, Mapping):
-                next_states.extend(current.values())
-        states = next_states
-        if not states:
-            break
-    return {index: tuple(values) for index, values in cardinalities.items()}
 
 
 def _validate_nested_collection_cardinality_payloads(
